requirements.txt
requests
dotenv
rich
google-generativeai
flask
beautifulsoup4